reorganized .py files into modules, documented modules in Layout of README.md
