+++ README.md
@@ -1,5 +1,36 @@
-Put all your good documentation here. 
+Welcome to Risknet! This is a downloadable Pip package where you can access and run an XGBoost pipeline.
 
+# Layout
+- src/risknet
+   - config: holds `conf.yaml`
+      - `conf.yaml`: helps with setup
+      - `handlers.py`
+   - data: empty folder where user can store FM dataset
+   - jobs
+      - `__init__.py`
+      - `cloud_etl.py`: helps save files to cloud
+   - proc: contains preprocessing steps like feature encoding, label prep, and train-test splits
+      - `__init__.py`
+      - `encoder.py`: feature engineering/encoding categoricals
+      - `label_prep.py`: defines default, progress on loan
+      - `reducer.py`: reduces features based on importance, also train/test/val splits
+   - run: contains files for running pipeline
+      - `__init__.py`
+      - `main.py`: currently empty. WIP define pipeline as a function and call here
+      - `model.py`: defines the model class + functions
+      - `pipeline.py`: calls functions to execute the pipeline
+   - sys: contains files to set up system environment and logging
+      - `__init__.py`
+      - `log.py`: sets up logger
+      - `managers.py`: sets up the dask manager
+      - `system.py`: defines creating and removing files via the `sys` package
+   -`init.py`
+   -`main.py`: logs start, stop time for running the program (including downloading packages from setup.cfg)
+   - tests: store tests here
+      -`__init__.py`
+      -`test_stub.py`: currently only asserts True == True. No tests added yet.
+
+# Steps to Update Version
 To update the version:
 - `rm -rf dist build` to remove build folder
 - manually remove "egg-info" folder. This will change `src` to `src/risknet`.
@@ -14,23 +45,36 @@ To update the version:
 
 # History:
 0.0.1: Ran into problems with installing pytest-cov
+
 0.0.2: Got error:
 `ERROR: Could not find a version that satisfies the requirement dask[complete] (from risknet) (from versions: none) ERROR: No matching distribution found for dask[complete]`
+
 0.0.3: Tried moving dask into [options.extras_require] not install_requires. Got error for flake8
+
 0.0.4: Commented out more packages.
 Got error: `error: package directory 'lib3/yaml' does not exist` and `metadata-generation-failed`.
+
 0.0.5: Moved types-YAML into options, still get `'lib3/yaml does not exist'` error :/
+
 0.0.6: we're removing YAML as a test
 Got error `ERROR: Could not find a version that satisfies the requirement typing (from risknet) (from versions: none) ERROR: No matching distribution found for typing`
+
 0.0.10: reverting to version 0.0.6, checking if it works --success! But still error `No matching distribution found for typing`
+
 0.0.11: what happens when we remove typing (since it's part of stdlib in python >3.5)?
+
 0.0.12: commented out all packages added by EC
+
 0.0.13: Got new error:
 `ERROR: Cannot install pandas==1.3.4 and risknet==0.0.12 because these package versions have conflicting dependencies.
+
 The conflict is caused by:
     risknet 0.0.12 depends on numpy
     pandas 1.3.4 depends on numpy>=1.17.3; platform_machine != "aarch64" and platform_machine != "arm64" and python_version < "3.10"`
+
 Solution: will try setting python > 3.10
+
+
 ### SUCCESS! Version 0.0.13 can download from emily's (base)
 Caveats:
 - Probably can only work because it has all dependencies already installed in the env (it threw an error when I tried to run it in risknet_test)
