+++ src/risknet/proc/encoder.py
