+++ src/risknet/proc/reducer.py
